+++ lib/service/Health.dart
@@ -55,8 +55,6 @@ class Health with Service implements NotificationsListener {
   static const String notifyCountyStatusAvailable   = "edu.illinois.rokwire.health.county.status.available";
   static const String notifyUpdatedHistoryAvailable = "edu.illinois.rokwire.health.updated.history.available";
   
-  static const String notifyHealthStatusChanged     = "edu.illinois.rokwire.health.health_status.changed";
-
   static const String _historyFileName              = "history.json";
 
   HealthUser _user;
@@ -69,6 +67,7 @@ class Health with Service implements NotificationsListener {
   File     _historyCacheFile;
   List<Covid19History> _historyCache;
   Map<String, HealthRulesSet2> _rulesCache;
+  Map<String, Map<String, dynamic>> _accessRulesCache;
 
   bool _processingCountyStatus;
   bool _loadingUpdatedHistory;
@@ -85,6 +84,7 @@ class Health with Service implements NotificationsListener {
 
   Health._internal() {
     _rulesCache = Map<String, HealthRulesSet2>();
+    _accessRulesCache =  Map<String, Map<String, dynamic>>();
   }
 
 
@@ -191,8 +191,7 @@ class Health with Service implements NotificationsListener {
       Response response = await Network().get(url, auth: NetworkAuth.User);
       if (response?.statusCode == 200) {
         Covid19Status status = await Covid19Status.decryptedFromJson(AppJson.decodeMap(response.body), _userPrivateKey);
-        _lastCovid19Status = status?.blob?.healthStatus;
-        _updateExposureReportTarget(status: status);
+        _applyLastCovid19Status(status);
         return status;
       }
     }
@@ -206,8 +205,7 @@ class Health with Service implements NotificationsListener {
       String post = AppJson.encode(encryptedStatus?.toJson());
       Response response = await Network().put(url, body: post, auth: NetworkAuth.User);
       if (response?.statusCode == 200) {
-        _lastCovid19Status = status?.blob?.healthStatus;
-        _updateExposureReportTarget(status: status);
+        _applyLastCovid19Status(status);
         return true;
       }
     }
@@ -900,15 +898,24 @@ class Health with Service implements NotificationsListener {
   }
 
   set _lastCovid19Status(String healthStatus) {
-    if (covid19HealthStatusIsValid(healthStatus) && (healthStatus != _lastCovid19Status)) {
+    Storage().lastHealthCovid19Status = healthStatus;
+  }
+
+  void _applyLastCovid19Status(Covid19Status status) {
+    String oldStatus = _lastCovid19Status;
+    String newStatus = status?.blob?.healthStatus;
+    if (covid19HealthStatusIsValid(newStatus) && (oldStatus != newStatus)) {
+      _lastCovid19Status = newStatus;
+
       Analytics().logHealth(
         action: Analytics.LogHealthStatusChangedAction,
-        status: healthStatus,
-        prevStatus: lastCovid19Status
+        status: newStatus,
+        prevStatus: oldStatus
       );
-      Storage().lastHealthCovid19Status = healthStatus;
-      NotificationService().notify(notifyHealthStatusChanged, null);
+
     }
+    _applyAccessFromStatus(status: status);
+    _updateExposureReportTarget(status: status);
   }
 
   void _updateExposureReportTarget({Covid19Status status}) {
@@ -1310,40 +1317,76 @@ class Health with Service implements NotificationsListener {
   // Consolidated Rules
 
   Future<HealthRulesSet2> loadRules2({String countyId, bool force}) async {
-    HealthRulesSet2 countyRules;
+    HealthRulesSet2 rules;
     if (countyId == null) {
       countyId = _currentCountyId;
     }
     if (countyId != null) {
-      countyRules = (force != true) ? _rulesCache[countyId] : null;
-      if (countyRules == null) {
+      rules = (force != true) ? _rulesCache[countyId] : null;
+      if (rules == null) {
         String appVersion = AppVersion.majorVersion(Config().appVersion, 2);
         String url = "${Config().healthUrl}/covid19/crules/county/$countyId";
         Response response = await Network().get(url, auth: NetworkAuth.App, headers: { Network.RokwireVersion : appVersion });
         String responseBody = (response?.statusCode == 200) ? response.body : null;
 //TMP:  String responseBody = await rootBundle.loadString('assets/sample.health.rules.json');
         Map<String, dynamic> responseJson = (responseBody != null) ? AppJson.decodeMap(responseBody) : null;
-        countyRules = (responseJson != null) ? HealthRulesSet2.fromJson(responseJson) : null;
-        if (countyRules != null) {
-          _rulesCache[countyId] = countyRules;
+        rules = (responseJson != null) ? HealthRulesSet2.fromJson(responseJson) : null;
+        if (rules != null) {
+          _rulesCache[countyId] = rules;
         }
       }
     }
-    return countyRules;
+    return rules;
   }
 
   // Access Rules
 
-  Future<Map<String, dynamic>> _loadAccessRules({String countyId}) async {
-    String url = "${Config().healthUrl}/covid19/access-rules/county/$countyId";
-    Response response = await Network().get(url, auth: NetworkAuth.App);
-    String responseBody = (response?.statusCode == 200) ? response.body : null;
-    return (responseBody != null) ? AppJson.decodeMap(responseBody) : null; 
+  Future<Map<String, dynamic>> _loadAccessRules({String countyId, bool force}) async {
+
+    Map<String, dynamic> accessRules;
+    if (countyId != null) {
+      accessRules = (force != true) ? _accessRulesCache[countyId] : null;
+      if (accessRules == null) {
+        String url = "${Config().healthUrl}/covid19/access-rules/county/$countyId";
+        Response response = await Network().get(url, auth: NetworkAuth.App);
+        String responseBody = (response?.statusCode == 200) ? response.body : null;
+        accessRules = (responseBody != null) ? AppJson.decodeMap(responseBody) : null; 
+        if (accessRules != null) {
+          _accessRulesCache[countyId] = accessRules;
+        }
+      }
+    }
+
+    return accessRules;
   }
 
   Future<bool> isAccessGranted(String healthStatus) async {
-    Map<String, dynamic> accessRules = await _loadAccessRules(countyId: _currentCountyId);
+    Map<String, dynamic> accessRules = await _loadAccessRules(countyId: _currentCountyId, force: true);
     return (accessRules != null) && (accessRules[healthStatus] == kCovid19AccessGranted);
+  }
+
+  Future<void> _applyAccessFromStatus({Covid19Status status}) async {
+    if (Config().settings['covid19ReportBuildingAccess'] == true) {
+      Map<String, dynamic> accessRules = await _loadAccessRules(countyId: _currentCountyId);
+      if (accessRules != null) {
+        String access = accessRules[status?.blob?.healthStatus];
+        if ((access != null) && (access != Storage().lastHealthCovid19Access)) {
+          if (await _logAccessUpdate(status.dateUtc, access)) {
+            Storage().lastHealthCovid19Access = access;
+          }
+        }
+      }
+    }
+  }
+
+  Future<bool> _logAccessUpdate(DateTime dateUtc, String access) async {
+    String url = "${Config().healthUrl}/covid19/building-access";
+    String post = AppJson.encode({
+      'date': healthDateTimeToString(dateUtc),
+      'access': access
+    });
+    Response response = await Network().put(url, body: post, auth: NetworkAuth.User);
+    return (response?.statusCode == 200);
   }