+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Log building access updates [#202](https://github.com/rokwire/safer-illinois-app/issues/202)
 
 ## [2.6.12] - 2020-09-24
 ### Changed